Introduced MSPKG_DIR and BSPPKG_DIR
Introduced MSPKG_DIR and BSPPKG_DIR to enable customization of the
location of the BSP packages as this is not delivered by MS and can be
located in different directory. Override BSPPKG_DIR to the location
where the BSP packages are kept. The default is the PKG Build directory
(PKGBLD_DIR).